chore: cleans the code and adds metadata
