+++ packages/examples/nodejs-interactive/src/RpcRequests.ts
@@ -1,35 +1,14 @@
 import { Buffer } from 'buffer';
-
-interface RpcRequest {
-  method: string;
-  params: any[];
-}
-
-const polygonDetails = {
-  chainId: '0x89',
-  chainName: 'Polygon',
-  blockExplorerUrls: ['https://polygonscan.com'],
-  nativeCurrency: { symbol: 'MATIC', decimals: 18 },
-  rpcUrls: ['https://polygon-rpc.com/'],
-}
-
-export const chains = {
-  GOERLI: '0x5',
-  SEPOLIA: '0xaa36a7',
-  MAINNET: '0x1',
-  POLYGON: '0x89',
-}
+import { RpcRequest, polygonDetails } from './Constants';
 
 export const sendTransactionRequest = (selectedAddress: string, amount: string = '0x5AF3107A4000'): RpcRequest => {
   const to = '0x0000000000000000000000000000000000000000';
   const transactionParameters = {
-    to, // Required except during contract publications.
-    from: selectedAddress, // must match user's active address.
+    to,
+    from: selectedAddress,
     value: amount,
   };
 
-    // txHash is a hex string
-    // As with any RPC call, it may throw an error
   return {
     method: 'eth_sendTransaction',
     params: [transactionParameters],
